[FE] [MODIFY] moved function into right file
