+++ packages/visitorWeb/src/services/userCalls.ts
@@ -1,6 +1,7 @@
 import axios from "axios";
 
 const baseUrl = `${process.env.DB_HOST}${process.env.DB_HOST_PORT}/api/login/`;
+const baseUrl1 = `${process.env.DB_HOST}${process.env.DB_HOST_PORT}/api/user/`;
 
 export const checkIfVisitorTokenIsValid = async (body: any) => {
   try {
@@ -20,5 +21,22 @@ export const checkIfVisitorTokenIsValid = async (body: any) => {
   } catch (error) {
     console.error("Error fetching the Users:", error);
     throw new Error("Error fetching the Users");
+  }
+};
+
+export const checkIfVisitorUserExist = async (body: any) => {
+  try {
+    const response = await axios({
+      method: "GET",
+      url: baseUrl1 + 'userVisitorExist',
+      params: body,
+      headers: {
+        "content-type": "application/json",
+      },
+    });
+    return response.data;
+  } catch (error) {
+    console.error("Error while checking visitor user:", error);
+    throw new Error("Error checking visitor user");
   }
 };
